Add check on length of extra atom or bond features
This check validates if the number of extra atom or
bond features matches the number of atoms or bonds
in the molecules and throws an error if this is not
the case. This check is done in data.py and not before
to avoid the computational cost of making additional
RDKit molecule graph structures.