Change imports in make_predictions.py to be v2.
Not sure of where featurizers are/occur so commented them out.
UncertaintyCalibrator is not yet implemented so is set to None.
The new dataloader is MolGraphDataLoader() instead of MoleculeDataLoader().
The models aren't just MoleculeModel(), they can be a range of model types.
Removed some inputs to UncertaintyEstimator because they are needed in the minimal implementation or are built into the model.