Add changes for features and atom descriptors CLI arguments
This commit makes some changes to help control the scaling of additional molecular features and atom_descriptors that may be concatenated to the representation after message passing. datasets.py have had normalize_inputs edited to check if the arrays of features are None. train.py is modified to apply the scaling to train, val, and test datasets and save the scalers. predict.py is modified to use the saved scaler to scale inputs during inference.