The SSL_pretrain and DDP version of chemprop
The SSL pretrain version of chemprop.
Here we have developed the cutomized masking and contrastive learning precedure for the directed graph of chemprop.
In addition, a large mpn encoder version is created for chemprop as well, but it will only be useful when it is in SSL_pretrain mode. This is aiming at not changing other functions.
Finally, the DDP version of training is enabled. This is because normally a SSL pretrain will have a huge amount of data, and to achieve a good contrastive learning performance, a huge batch size is needed. To ensure it can train fast and have huge batch size, the DDP train version is realized.

The code has passed all the tests, with two additional test, 1. test the SSL pretrain 2. test the DDP SSL pretrain.

Thanks!

Co-Authored-By: Kevin Greenman <35846516+kevingreenman@users.noreply.github.com>
Co-Authored-By: shomikverma <40326897+shomikverma@users.noreply.github.com>