Adapt CGR unit tests for __call__ to the update of torch_scatter
The introduction of torch_scatter changes the attribute of MolGraph to E, V, edge_index, and rev_edge_index. Others are gone. The Test of V is mainly covered by the test of _calc_node_feature_matrix; the test of E should be partially covered by the test of _calc_edge_feature. So the test of __call__ majorly covers 1. the shape of E and V; 2. examine edge_index and rev_edge_index are expected. When testing the later two attributes, an equivalent but slightly different implementation is used.