Initial Draft of `--canonicalize-smiles`
This is a quick draft of the new flag discussed in https://github.com/chemprop/chemprop/issues/1077 - probably doesn't run, but will start the discussion.

I've just added an optional (for backwards compatibility) argument throughout the codebase which controls whether or not we call `rdkit.Chem.CanonSmiles` before running `MolFromSmiles` inside `make_mol`.

A better solution would be to canonicalize the input smiles and then save them to the output files next to their original versions, but the implementation here is way simpler (perhaps MVP is best?).

Comments welcome!