Add explicit version lists for chemprop 2.x releases
External dependencies occasionally introduce backwards incompatible changes in minor versions, such as when torch changed weights_only to true by default, breaking chemprop 2.1.0 when installed from scratch (becase we had developed 2.1.0 before this change was made).

This PR provides a basic fix to this problem - all of the dependency versions from the Docker files (which captures a _known working_ set of dependencies) are now provided in the repo.

Resolves https://github.com/chemprop/chemprop/issues/1197